@@ -2,25 +2,25 @@ require 'sidekiq/web'
 
 
 Upandsell::Application.routes.draw do
-  devise_for :users, controllers: { confirmations: 'confirmations',
-    registrations: "registrations" }
-    if Rails.env.development?
-      mount MailPreview => 'mail_view'
-    end
-    root 'landing#index'
-    get 'pricing' => 'site#pricing'
-    get 'privacy' => 'site#privacy'
-    get 'terms' => 'site#terms'
-    get 'unsubscribe/u/:user/:type/:signature' => 'emails#unsubscribe_user',
-    :as => 'unsubscribe_user'
-    get 'confirm_unsubscribe/u/:user/:type/:signature' => 'emails#confirm_unsubscribe_user',
-    :as => 'confirm_unsubscribe_user'
+  devise_for :users, controllers: { confirmations: 'confirmations', registrations: "registrations" }
 
-    get 'unsubscribe/o/:order/:signature' => 'emails#unsubscribe_order',
-    :as => 'unsubscribe_order'
-    get 'confirm_unsubscribe/o/:order/:signature' => 'emails#confirm_unsubscribe_order',
-    :as => 'confirm_unsubscribe_order'
-    get 'unsubscribed' => 'emails#unsubscribed'
+  get '/auth/:provider/callback' => 'integrations#create'
+
+ # Front-end
+ root 'landing#index'
+ get 'pricing' => 'site#pricing'
+ get 'privacy' => 'site#privacy'
+ get 'terms' => 'site#terms'
+ get 'unsubscribe/u/:user/:type/:signature' => 'emails#unsubscribe_user',
+ :as => 'unsubscribe_user'
+ get 'confirm_unsubscribe/u/:user/:type/:signature' => 'emails#confirm_unsubscribe_user',
+ :as => 'confirm_unsubscribe_user'
+
+ get 'unsubscribe/o/:order/:signature' => 'emails#unsubscribe_order',
+ :as => 'unsubscribe_order'
+ get 'confirm_unsubscribe/o/:order/:signature' => 'emails#confirm_unsubscribe_order',
+ :as => 'confirm_unsubscribe_order'
+ get 'unsubscribed' => 'emails#unsubscribed'
 
   #product page
   get '/p/:slug' => 'products#show', :as => 'product_slug'
