@@ -1,6 +1,4 @@
-  # app/mailers/mail_preview.rb or lib/mail_preview.rb
-  class MailPreview < MailView
-    # Pull data from existing fixtures
+class MailPreview < ActionMailer::Preview
     def sold_email
       user = User.first
       order = Order.first
@@ -18,11 +16,5 @@
       user = User.first
       mail = UserMailer.welcome_email(user)
       mail
-    end
-
-    # Stub-like
-    def forgot_password
-      user = Struct.new(:email, :name).new('name@example.com', 'Jill Smith')
-      mail = UserMailer.forgot_password(user)
     end
   end
